Update README.md
The need for extended battery life is highlighted by the growing use of lithium-ion batteries in daily
life. In this context, battery endurance, or how long a pack or cell can operate without needing to be
recharged, is referred to as battery life. While there are different ways to enhance battery life (more
efficient parts or batteries with greater capacity), there is also a higher expectation for the state of
charge (SOC) prediction's accuracy in these applications. Additionally, for bigger battery systems,
such as those found in electric vehicles, there is a desire for greater longevity and even second life
applications. The monitoring of state of the battery’s condition will be discussed in this thesis.
Electrochemical impedance spectroscopy (EIS), broadband impedance spectroscopy (BIS), and
pseudo-random sequence (PRS) injections are the three main methods used in the field of battery
state monitoring. They are all aimed towards recording a cell's distinctive complicated impedance
curve and determining the cell's health based on how the curve behaves over time.